DebugLevel to DL, update generic_readers
move some code originally from first commit to generic_readers.

Force push for clarity
